Crate rsp2_fs_util

Structs

ActualTempDir
A directory in the filesystem that is automatically deleted when it goes out of scope.

Copy
Builder for configuring a cp -a command.

Move
Builder for configuring a mv command.

TempDir
Wrapper around tempdir::TempDir that does not destroy the directory on unwind.

Enums

FsError

Functions

canonicalize
Wrapper around std::fs::canonicalize that adds context.

canonicalize_parent
Canonicalizes a path where the final component need not exist.

copy
Wrapper around std::fs::copy that adds context.

cp_a
Calls cp -aT.

create
Wrapper around File::create that adds context.

create_dir
Wrapper around std::fs::create_dir that adds context.

hard_link

mv
Move a file or directory, possibly across filesystems.

open
Wrapper around File::open that adds context.

open_text
Wrapper around File::open that adds context and makes a BufReader.

remove_dir
Wrapper around std::fs::remove_file that adds context.

remove_file
Wrapper around std::fs::remove_file that adds context.

rm_rf
Simulates rm -rf.

write
Wrapper around std::fs::write that adds context.

Type Definitions

FsResult